Refactor codegen into more pieces
